requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.3.1
ckzg==2.1.1
cytoolz==1.0.1
Django==5.2
django-cors-headers==4.7.0
django-decouple==2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
gunicorn==23.0.0
hexbytes==1.3.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
parsimonious==0.10.0
psycopg2-binary==2.9.10
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pymongo==3.11.4
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rlp==4.1.0
rpds-py==0.24.0
sqlparse==0.5.3
toolz==1.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
